requirements/base.txt
aiocontextvars==0.2.2
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.10.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
environs==9.3.1
exceptiongroup==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytest==7.2.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvloop==0.17.0
yarl==1.8.2
